fix: added handling of `files_added` event on read-only boards
