+++ Makefile
@@ -41,7 +41,7 @@ messages:
 
 
 env:
-	python3 -m venv env
+	python3 -m venv ENV
 	ENV/bin/python -m pip install -U pip wheel setuptools
 	ENV/bin/python -m pip install -r requirements.txt
