+++ devscripts/runserver.sh
@@ -4,8 +4,9 @@ source $BASEDIR/init.sh
 
 $BASEDIR/make_static.sh
 
-echo; echo $(colored $CYAN "migrating database... "); echo
-gosu moodle ENV/bin/python manage.py migrate --settings=local_settings
+#echo; echo $(colored $CYAN "migrating database... "); echo
+#gosu moodle ENV/bin/python manage.py migrate --settings=local_settings
 
 echo;echo $(colored $CYAN "starting development server:"); echo
-gosu moodle ENV/bin/python manage.py runserver --settings=local_settings 127.0.0.1:8001
+#gosu moodle ENV/bin/python manage.py runserver --settings=local_settings 127.0.0.1:8001
+ENV/bin/python manage.py runserver
